Updated data loading so that it works on my laptop. Some additions to plotting.

+++ utils/other.py
@@ -195,7 +195,7 @@ def tile_raster_images(X, img_shape, tile_shape, tile_spacing=(0, 0),
                                  dtype='uint8')
         else:
             out_array = np.zeros((out_shape[0], out_shape[1], 4),
-                                 dtype=X.dtype)
+                                 dtype=X[0].dtype)
 
         # colors default to 0, alpha defaults to 1 (opaque)
         if output_pixel_vals:
